Implement finder for notification counters

+++ codebrag-dao/src/main/scala/com/softwaremill/codebrag/dao/reporting/views/NotificationCountersView.scala
@@ -1,3 +1,3 @@
 package com.softwaremill.codebrag.dao.reporting.views
 
-case class NotificationCountersView(pendingCommitCount: Int, followupCount: Int)
+case class NotificationCountersView(pendingCommitCount: Long, followupCount: Long)
